+++ examples/01-choco/auth.tf
@@ -7,13 +7,32 @@ provider "azurerm" {
   }
 }
 
+provider "azapi" {
+}
+
+provider "oci" {
+  tenancy_ocid = var.TENANCY_OCID
+  user_ocid    = var.USER_OCID
+  private_key  = var.PRIVATE_KEY
+  fingerprint  = var.FINGERPRINT
+  region       = "eu-frankfurt-1"
+}
+
 #Versions
 terraform {
-  required_version = ">=1.3.6"
+  required_version = ">= 1.6.0"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
       version = "3.99.0"
+    }
+    azapi = {
+      source  = "azure/azapi"
+      version = ">= 1.9.0"
+    }
+    oci = {
+      source  = "oracle/oci"
+      version = "5.38.0"
     }
   }
 }