+++ examples/01-choco/context_azure_dev.tf
@@ -16,6 +16,9 @@ module "context_choco_azure_dev" {
   custom_tags = {
     HQ = "Brussels"
   }
+
+  #Regions
+  oci_regions = data.oci_identity_regions.regions.regions
 }
 output "azure_dev_context" {
   value = module.context_choco_azure_dev.context