+++ examples/01-choco/outputs.tf
@@ -1,0 +1,3 @@
+output "oci_short_regions" {
+  value = module.context_choco_azure_dev.oci_short_regions
+}