+++ examples/02-candy/auth.tf
@@ -7,6 +7,14 @@ provider "azurerm" {
   }
 }
 
+provider "oci" {
+  tenancy_ocid = var.TENANCY_OCID
+  user_ocid    = var.USER_OCID
+  private_key  = var.PRIVATE_KEY
+  fingerprint  = var.FINGERPRINT
+  region       = "eu-frankfurt-1"
+}
+
 #Versions
 terraform {
   required_version = ">=1.3.6"