+++ modules.tf
@@ -4,20 +4,24 @@
 
 #Brussels Chocolate Factory
 module "choco" {
-  count       = local.subsidiary == "choco" ? 1 : 0
-  source      = "./modules/choco"
-  cloud       = local.cloud
-  environment = local.environment
-  region      = local.region
+  count         = local.subsidiary == "choco" ? 1 : 0
+  source        = "./modules/choco"
+  cloud         = local.cloud
+  environment   = local.environment
+  region        = local.region
+  azure_regions = local.azure_short_regions
+  oci_regions   = local.oci_short_regions
 }
 
 #Cologne Candy Shop
 module "candy" {
-  count       = local.subsidiary == "candy" ? 1 : 0
-  source      = "./modules/candy"
-  cloud       = local.cloud
-  environment = local.environment
-  region      = local.region
+  count         = local.subsidiary == "candy" ? 1 : 0
+  source        = "./modules/candy"
+  cloud         = local.cloud
+  environment   = local.environment
+  region        = local.region
+  azure_regions = local.azure_short_regions
+  oci_regions   = local.oci_short_regions
 }
 
 #Istanbul Icecream Industries