+++ modules/candy/main.tf
@@ -1,11 +1,11 @@
-#  $$$$$$\  $$\   $$\  $$$$$$\   $$$$$$\   $$$$$$\  
-# $$  __$$\ $$ |  $$ |$$  __$$\ $$  __$$\ $$  __$$\ 
-# $$ /  \__|$$ |  $$ |$$ /  $$ |$$ /  \__|$$ /  $$ |
-# $$ |      $$$$$$$$ |$$ |  $$ |$$ |      $$ |  $$ |
-# $$ |      $$  __$$ |$$ |  $$ |$$ |      $$ |  $$ |
-# $$ |  $$\ $$ |  $$ |$$ |  $$ |$$ |  $$\ $$ |  $$ |
-# \$$$$$$  |$$ |  $$ | $$$$$$  |\$$$$$$  | $$$$$$  |
-#  \______/ \__|  \__| \______/  \______/  \______/ 
+#  $$$$$$\   $$$$$$\  $$\   $$\ $$$$$$$\ $$\     $$\ 
+# $$  __$$\ $$  __$$\ $$$\  $$ |$$  __$$\\$$\   $$  |
+# $$ /  \__|$$ /  $$ |$$$$\ $$ |$$ |  $$ |\$$\ $$  / 
+# $$ |      $$$$$$$$ |$$ $$\$$ |$$ |  $$ | \$$$$  /  
+# $$ |      $$  __$$ |$$ \$$$$ |$$ |  $$ |  \$$  /   
+# $$ |  $$\ $$ |  $$ |$$ |\$$$ |$$ |  $$ |   $$ |    
+# \$$$$$$  |$$ |  $$ |$$ | \$$ |$$$$$$$  |   $$ |    
+#  \______/ \__|  \__|\__|  \__|\_______/    \__|    
 
 locals {
   #Input Parameters
@@ -16,9 +16,11 @@ locals {
   #Context
   context = merge(
     {
-      cloud = local.cloud,
-      environment = local.environment,
-      region = local.region,
+      cloud         = local.cloud,
+      environment   = local.environment,
+      region        = local.region,
+      cloud_regions = local.cloud_regions,
+      short_region  = local.short_region,
     },
     local.global,
     {
@@ -34,7 +36,7 @@ locals {
   )
 
   #Direct Data Access
-  azure  = merge(local.azure_global, local.azure_env)
-  oci    = merge(local.oci_global, local.oci_env)
+  azure = merge(local.azure_global, local.azure_env)
+  oci   = merge(local.oci_global, local.oci_env)
 
 }