+++ variables.tf
@@ -31,36 +31,6 @@ variable "environment" {
   description = "Environment - place in the SDLC lifecycle"
 }
 
-# variable "environment_azure" {
-#   type        = string
-#   description = "Azure Environment - place in the SDLC lifecycle"
-#   default     = null
-# }
-
-# variable "environment_aws" {
-#   type        = string
-#   description = "AWS Environment - place in the SDLC lifecycle"
-#   default     = null
-# }
-
-# variable "environment_gcp" {
-#   type        = string
-#   description = "GCP Environment - place in the SDLC lifecycle"
-#   default     = null
-# }
-
-# variable "environment_oci" {
-#   type        = string
-#   description = "OCI Environment - place in the SDLC lifecycle"
-#   default     = null
-# }
-
-# variable "environment_onprem" {
-#   type        = string
-#   description = "OnPrem Environment - place in the SDLC lifecycle"
-#   default     = null
-# }
-
 #Region
 variable "region" {
   type        = string
@@ -105,4 +75,60 @@ variable "custom_tags" {
   type        = map(any)
   default     = {}
   description = "Custom Resource tags"
+}
+
+# $$$$$$$\  $$$$$$$$\  $$$$$$\  $$$$$$\  $$$$$$\  $$\   $$\  $$$$$$\  
+# $$  __$$\ $$  _____|$$  __$$\ \_$$  _|$$  __$$\ $$$\  $$ |$$  __$$\ 
+# $$ |  $$ |$$ |      $$ /  \__|  $$ |  $$ /  $$ |$$$$\ $$ |$$ /  \__|
+# $$$$$$$  |$$$$$\    $$ |$$$$\   $$ |  $$ |  $$ |$$ $$\$$ |\$$$$$$\  
+# $$  __$$< $$  __|   $$ |\_$$ |  $$ |  $$ |  $$ |$$ \$$$$ | \____$$\ 
+# $$ |  $$ |$$ |      $$ |  $$ |  $$ |  $$ |  $$ |$$ |\$$$ |$$\   $$ |
+# $$ |  $$ |$$$$$$$$\ \$$$$$$  |$$$$$$\  $$$$$$  |$$ | \$$ |\$$$$$$  |
+# \__|  \__|\________| \______/ \______| \______/ \__|  \__| \______/ 
+
+variable "azure_region_recommended_only" {
+  type        = bool
+  default     = true
+  description = "(Optional) Only return recommended regions for Azure."
+}
+
+variable "azure_regions" {
+  type        = any
+  default     = null
+  description = <<EOT
+  You can use this variable to dynamically populate data for all available regions in Azure.
+
+  You can use the https://github.com/Azure/terraform-azurerm-regions module with the Azure/azapi provider:
+
+  ```hcl
+  module "regions" {
+    source  = "Azure/regions/azurerm"
+    version = "0.6.0"
+  }
+  ```
+
+  And feed the output of the module to the context module:
+
+  ```hcl
+  module "context" {
+    azure_regions = module.regions.regions
+  }
+  ```
+  EOT
+}
+
+variable "oci_regions" {
+  type        = any
+  default     = null
+  description = <<EOT
+  You can use this variable to dynamically define the region short codes for all available regions in OCI.
+
+  ```hcl
+  data "oci_identity_regions" "regions" {}
+
+  module "context" {
+    oci_regions = data.oci_identity_regions.regions.regions
+  }
+  ```
+  EOT
 }